+++ .env
@@ -1,0 +1,2 @@
+NEXT_DATOCMS_API_TOKEN=d084e9f8cb6f25a046791851c77e65
+MAPS_API_KEY=AIzaSyCLo1g8znFtrQg-UovgeyOxIH4ZqzkddHI
