+++ components/content.js
@@ -3,8 +3,6 @@ import styled from "styled-components";
 export default styled.div`
   display: flex;
   flex-direction: column;
-  max-width: 1000px;
-  margin-left: auto;
-  margin-right: auto;
-  padding: 20px;
+  margin-top: 48px;
+  margin-bottom: 48px;
 `;
