+++ package.json
@@ -10,18 +10,16 @@
   },
   "dependencies": {
     "babel-plugin-styled-components": "^1.12.0",
-    "google-map-react": "^2.1.9",
     "graphql": "^15.5.0",
     "graphql-request": "^3.4.0",
     "hamburger-react": "^2.4.0",
     "html-react-parser": "^1.2.4",
-    "next": "10.0.6",
+    "next": "10.0.9",
     "next-sitemap": "^1.6.9",
-    "react": "17.0.1",
-    "react-datocms": "^1.5.5",
-    "react-dom": "17.0.1",
-    "sass": "^1.32.8",
+    "react": "17.0.2",
+    "react-datocms": "^1.5.6",
+    "react-dom": "17.0.2",
     "styled-components": "^5.2.1",
-    "styled-icons": "^10.29.0"
+    "styled-icons": "^10.31.0"
   }
 }
