+++ pages/_app.js
@@ -1,13 +1,31 @@
-import { ThemeProvider, createGlobalStyle } from "styled-components";
+import { createGlobalStyle, ThemeProvider } from "styled-components";
 import Router from "next/router";
 import * as gtag from "../lib/gtag";
 
+// Notice how we track pageview when route is changed
 Router.events.on("routeChangeComplete", (url) => gtag.pageview(url));
 
 const GlobalStyle = createGlobalStyle`
 
 * {
   box-sizing: border-box;
+}
+
+html {
+  height: -webkit-fill-available;
+}
+
+body {
+  min-height: 100vh;
+  /* mobile viewport bug fix */
+  min-height: -webkit-fill-available;
+}
+
+#__next {
+  min-height: 100vh;
+  min-height: -webkit-fill-available;
+  display: flex;
+  flex-direction: column;
 }
 
 html,
@@ -18,11 +36,14 @@ body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
     Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-    font-family: quasimoda, sans-serif;
-    color: #272727;
-    background-color: #f8f8ff;
+  font-family: quasimoda, sans-serif;
+  color: #272727;
+  background-color: #f8f8ff;
 }
 
+main {
+  flex: 1;
+}
 .trajan {
   font-family: trajan-pro-3, serif;
 }
@@ -30,16 +51,13 @@ body {
 a {
   color: inherit;
   text-decoration: none;
+  width: fit-content;
+  height: fit-content;
 }
 
-p, span {
-  font-size: 20px;
-  margin-top: 52px;
-  margin-bottom: 52px;
-}
-
-h1, h2, h3, h4, h5, h6 {
-  margin-top: 52px;
+span {
+  margin-top: 6px;
+  margin-bottom: 6px;
 }
 
 .hamburger-react {
@@ -48,15 +66,18 @@ h1, h2, h3, h4, h5, h6 {
     display: block;
     position: absolute!important;
     right: 20px;
-    top: 25px;
-    z-index: 2;
+    top: 40px;
+    z-index: 3;
   }
-}`;
+}
+`;
 
 const theme = {
-  background: "#f8f8ff",
-  text: "#272727",
-  brandBlue: "#158ba8",
+  colors: {
+    background: "#f8f8ff",
+    text: "#272727",
+    brandBlue: "#158ba8",
+  },
 };
 
 function MyApp({ Component, pageProps }) {
